solved problems in UserControllerTest and UserAuthControllerTest
