+++ src/test/java/com/alkemy/ong/controller/UserControllerTest.java
@@ -11,6 +11,7 @@ import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.web.servlet.MockMvc;
@@ -22,6 +23,9 @@ import java.util.List;
 import java.util.Optional;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.when;
+import static org.springframework.http.MediaType.APPLICATION_JSON;
+import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -40,6 +44,7 @@ class UserControllerTest {
     @Autowired
     UserMapper userMapper;
     List<UserEntity> userEntityList = new ArrayList<>();
+    List<UserDto> userDtoList = new ArrayList<>();
     @BeforeEach
     void setUp() {
         mockMvc = MockMvcBuilders.webAppContextSetup(context)
@@ -50,8 +55,12 @@ class UserControllerTest {
         userEntityList.add(user);
         userEntityList.add(user1);
 
+        UserDto userDto = userMapper.convertUserToDto(user);
+        userDtoList.add(userDto);
+
     }
     @Test
+    @WithMockUser(roles = "ADMIN")
     void deleteUser() throws Exception {
         UserEntity user = userEntityList.get(0);
         Long userId = user.getId();
@@ -61,10 +70,31 @@ class UserControllerTest {
                 .andExpect(status().is(204));
     }
     @Test
+    @WithMockUser(roles = "ADMIN")
     void shouldReturn404WhenDeletingNonExistingUser() throws Exception {
         Long userId = 33L;
         given(userService.findUserById(userId)).willReturn(Optional.empty());
         this.mockMvc.perform(delete("/users/{id}", userId))
                 .andExpect(status().is(204));
+    }
+
+    @Test
+    @WithMockUser(roles = "ADMIN")
+    void findUserById() throws Exception {
+        Long userId = 1L;
+        UserEntity user = new UserEntity(userId, false, "Andres", "Rodriguez", "andres@Email", "123456", "photo", null, null, null);
+        given(userService.findUserById(userId)).willReturn(Optional.of(user));
+        this.mockMvc.perform(get("/users/{id}", userId))
+                .andExpect(status().isOk());
+    }
+
+    @Test
+    @WithMockUser(roles = "ADMIN")
+    void getAllUsersD()  throws Exception {
+        when(userService.getAllUsers()).thenReturn(userDtoList);
+        mockMvc.perform(get("/users")
+                        .contentType(APPLICATION_JSON)
+                        .with(csrf()))
+                .andExpect(status().isOk());
     }
 }
